Move QuickBalance to commands.QuickBalance
