+++ server/Ion/src/main/kotlin/net/horizonsend/ion/server/Ion.kt
@@ -1,5 +1,6 @@
 package net.horizonsend.ion.server
 
+import net.horizonsend.ion.server.commands.QuickBalance
 import net.horizonsend.ion.server.listeners.dynmap.DynmapEnabledListener
 import net.horizonsend.ion.server.listeners.paper.Tweaks
 import net.starlegacy.PLUGIN
