+++ server/StarLegacy/src/main/kotlin/net/starlegacy/feature/starship/Interdiction.kt
@@ -1,10 +1,7 @@
 package net.starlegacy.feature.starship
 
-import net.horizonsend.ion.server.QuickBalance
+import net.horizonsend.ion.server.commands.QuickBalance
 import net.starlegacy.SLComponent
-import net.starlegacy.cache.nations.PlayerCache
-import net.starlegacy.cache.nations.RelationCache
-import net.starlegacy.database.schema.nations.NationRelation
 import net.starlegacy.feature.misc.CustomItems
 import net.starlegacy.feature.multiblock.Multiblocks
 import net.starlegacy.feature.multiblock.gravitywell.GravityWellMultiblock
