+++ server/StarLegacy/src/main/kotlin/net/starlegacy/feature/starship/subsystem/weapon/primary/PointDefenseSubsystem.kt
@@ -1,6 +1,6 @@
 package net.starlegacy.feature.starship.subsystem.weapon.primary
 
-import net.horizonsend.ion.server.QuickBalance
+import net.horizonsend.ion.server.commands.QuickBalance
 import kotlin.math.sqrt
 import net.starlegacy.feature.starship.active.ActivePlayerStarship
 import net.starlegacy.feature.starship.active.ActiveStarship
