+++ server/StarLegacy/src/main/kotlin/net/starlegacy/feature/starship/subsystem/weapon/secondary/TorpedoWeaponSubsystem.kt
@@ -1,7 +1,6 @@
 package net.starlegacy.feature.starship.subsystem.weapon.secondary
 
-import net.horizonsend.ion.server.QuickBalance
-import net.horizonsend.ion.server.QuickBalance.getBalancedValue
+import net.horizonsend.ion.server.commands.QuickBalance.getBalancedValue
 import java.util.concurrent.TimeUnit
 import net.starlegacy.feature.starship.active.ActiveStarship
 import net.starlegacy.feature.starship.subsystem.weapon.TargetTrackingCannonWeaponSubsystem
